+++ cockpit/lib/i18n/strings.g.dart
@@ -4,9 +4,9 @@
 /// To regenerate, run: `dart run slang`
 ///
 /// Locales: 3
-/// Strings: 3012 (1004 per locale)
+/// Strings: 3036 (1012 per locale)
 ///
-/// Built on 2026-09-23 at 20:42 UTC
+/// Built on 2026-09-26 at 02:41 UTC
 
 // coverage:ignore-file
 // ignore_for_file: type=lint, unused_import